update add liquidity modal ui
